Fix some issues with locals (in Parser + Machine)
The parser was not correctly processing the section declaring the types
of locals while parsing CodeSegment.  It had been using the count of how
many adjacent locals were of the same type as a single local with that
count as its default value.  This was causing machine to undercount how
many locals existed.  It also led to erroneously initializing some locals
to some unlikely initial values.

In setting up StackFrame it was only using locals being passed in to
initialize locals.  The problem here was that the incoming parameters
are actually the first locals from a Machine perspective.  So the fix
was to assign all parameters as first locals and then add the locals
declared in CodeSegment after those.  As a result this removes all
need to check for missing locals.  That in turn allows the use of a
single primitive array.

There is one outstanding question of how to initialize (zero) out
a local for V128.  I added a TODO and guard against trying to
zero it.  I expect there is some crash potential here since webasm
says all locals must be initialized but it is probably not crashing
due to lack of coverage for V128.