Make FunctionBody more final
In my last PR I recognized I made localTypes as a final primitive array
and thought it looked a bit weird that instructions were also not a
final primitive array.  Unless there is a need to mutably update
instructions this feels like it shows intent that this is a set in
stone definition as an artifact of parsing.

I also moved the unused ast into Ast and renamed it.  The benefit of
this new definition is it is uncoupled from FunctionBody and it obviously
already needs to know what Instructions are.  This may just be dead code
though?

Not sure if this is needed but I figured I may as well tie up what looked
like a loose end to me.