Builder helper class for HostImports (#428)
Hi all,

I'm getting familiar with this repository and wanted to start by
contributing a very small improvement.

In this case, I'm tackling #357, since @andreaTP said it could be a good
way to start.

Here I'm adding a `HostImportsBuilder` class that can help with the
situation described in #357.

If this is OK and we merge it, further refinements can be done in the
HostImports class to remove redundant constructors. Similar tactics
could be used for other classes too.

(might fix #357)