refactor: remove annoying obsolete log (#434)
System.out.println("Skipping Section with ID due to configuration: " +
sectionId);
Is a public log that we cannot hide, so it would be optimal to use the
provided logger to log it.

That set aside, I belive firmly that the provided logger should be
optional and nullable. If you disagree with me, I can return the logger
to be nonNull, but the sysout should be changed to logger.