Run Chicory tests on Android (#743)
This change adds a new Android project to run Chicory tests on Android
devices.

Android doesn't have official maven support, hence this is not created
as a maven project. Instead, this Android project has the ability to
compile the main project and copy its tests to run the as Android tests.

See the android-tests/README.md for details on how it works.

This first change only adds the runtime tests but similar setup can be
created for other modules.

Furhermore, currently only API 33+ is supported due to
InputStream.readAllBytes API usages, which is only added on API 33 to
Android.

The CI is setup to run tests on 33 and 35. It is kept as a separate
workflow file to avoid interference with the main build.

Spotless is setup with the KOTLINLANG default style, as there is no
matching AOSP style for ktfmt.

Bug: #535
Test: CI

---------

Co-authored-by: Edoardo Vacchi <evacchi@users.noreply.github.com>
Co-authored-by: Andrea Peruffo <andrea.peruffo1982@gmail.com>