Easy HostModule rename mechanism (#792)
This adds a new generated overload where the module name can be supplied
explicitly

```java
    public static HostFunction[] toHostFunctions(Simple functions, String moduleName) {
```

Looks like there are Wasm modules where the wasi function module is
called `wasi_unstable`.

This change is backward-compatible and provides an easy mechanism for
changing the module name to all the functions defined in a "HostModule".

cc. @chirino