WasmModuleInterface Exports is now a class with cached fields (#811)
This is a breaking change of the code generated by the
`WasmModuleInterface` annotation.

Also fixes: #725 

Currently `WasmModuleInterface` generates for the Wasm Module "Export"
section a Java `interface`, this is sub-optimal as we cannot "cache" the
direct exports on the module and we have to go, every time, through a
`Map<String, ..>` lookup.

With this change we propose to generate a full fledged `class` that can
be instantiated based on a `runtime.Instance` and will keep highly
optimized and Java idiomatic accesses to the exports.

To review this change I'd encourage reviewers to look at the folders
`host-module/it/target/it/**/target/generated-test-sources/test-annotations`
before and after the change.

The new encoding is much closer to what has been done manually for
real-world integrations in
[SQLite4j](https://github.com/roastedroot/sqlite4j/blob/main/src/main/java/io/roastedroot/sqlite4j/core/wasm/WasmDBExports.java)
and
[proxy-wasm-java-host](https://github.com/roastedroot/proxy-wasm-java-host/blob/main/proxy-wasm-java-host/src/main/java/io/roastedroot/proxywasm/ABI.java)
(cc. @chirino).

As a demo I also applied the proposed changes [to
Prism](https://github.com/ruby/prism/commit/5c7d25e13b459e33ca38c941140b15937bccf870)
and it seems to be a step in the right direction.