[proposals] Exception Handling  - initial implementation (#794)
This is an initial implementation of the Exception Handling proposal.
There are rough edges and a few details to be cleared out, but I'm
opening early to receive comments.

In particular a review by @SoniEx2 would be appreciated as they
implemented it in Wabt. :pray:

Useful links:
- https://webassembly.github.io/exception-handling/
-
https://github.com/WebAssembly/exception-handling/blob/main/proposals/exception-handling/Exceptions.md

I'll comment inline more this implementation tomorrow.