wasm-tools wat2wasm utility (#840)
There is a mysterious bug happening only in compiled mode when specific
versions of Java are used, this PR adds another way to use `wat2wasm`
with the interpreter mode.

To unblock #833 @chirino you can now use this one (please keep the
numbers low as it's taking forever to finish ~2 minutes).