workaround[wasi/quarkus]: Error: Detected an instance of Random/SplittableRandom class in the image heap
A Quarkus native app that creating an Instance using Wasi at build time was getting this error:
`Quarkus Native Build fails with: Error: Detected an instance of Random/SplittableRandom class in the image heap.`


This change the user to clear the random instance used in wasi by calling WasiPreview1.setRandom(null).  After the instance is initialized to avoid this error.


Fixes #853