Switch default to ThreadLocalRandom in WasiOpts (#855)
Fix for: #853

It took forever to figure out the best solution.
Handling of `ThreadLocalRandom` is already baked in GraalVM so we can
safely leverage it.

note to self:
I have no clue why the relevant substitutions for all the standard
library `Random` implementation are not provided by default. No one
wants a Random generator that is seeded at compile time ...