Nightly Zig testsuite (#884)
Opening for visibility, if anyone is interested in picking up the
challenge to fix the Symlink support of our Wasi layer
:slightly_smiling_face:

results:

https://github.com/andreaTP/chicory/actions/runs/14913567382/job/41893594463#step:5:6908

> 2677 passed; 118 skipped; 8 failed.

all of the failing tests are related to links/symlinks on the
FileSystem.

The patch is used to fix the printing to standard out and skip a few
additional tests.