Update the documentation to streamline the compiler (#916)
With those changes we remove both the compiler and the annotation
processors from the experimental phase.

AFAICT there are no outstanding issues to prevent doing it, I
added/moved the relevant documentation.

I think it's ok to leave the website a little ahead for a little time
while we rename the packages and roll out the next release.

---------

Co-authored-by: Edoardo Vacchi <evacchi@users.noreply.github.com>