Fix tail call optimization bug in the interpreter (#942)
@evacchi is implementing Tail Call Optimization in wazero, and suggested
to test SQLite as they found bugs there.
The compiler works fine, but the interpreter crashes due to a bug, you
can reproduce with this diff:


https://github.com/roastedroot/sqlite4j/compare/main...andreaTP:sqlite4j:check-tail-call

I'm not 100% confident about the implementation, as there are many
interpolations that are still working, but with those changes the
testsuite passes.
No issues found with the compiler :+1:

---------

Signed-off-by: Edoardo Vacchi <evacchi@users.noreply.github.com>
Co-authored-by: Edoardo Vacchi <evacchi@users.noreply.github.com>