compiler: externalize class collector logic (#980)
This PR adds a `ClassCollector` (for lack of a better name...). A class
collector:

- collects tuples `className, bytes` 
- stores the "main" class name (the Machine)
- exposes a "resolve" method to resolve `Class` to bytes

Note: this class really is exposing 2 things (and I am open to split
them -- they were originally 2 interfaces), one is the real collector,
the other is a resolver.

- The Collector is just rationalizing the behavior we already have in
the compiler to add all the classes we generate to a collection, _but_
it also, and most importantly **allows to plug a different
implementation that _does not_ perform class loading**.
  The reasons are
1) we don't really need to perform class loading in the maven compiler,
we only need to dump the classes onto disk
   2) we might need/want to perform class loading at a different stage 
3) on some platforms class loading might not be contextually feasible
as-is (for instance, if there are strict class loading policies in
effect)
- MachineFactory creation is now moved from the compiler into the
ClassLoadingCollector, with all the classloading-related logic.

- The resolver (`byte[] resolve(Class)`) is also driven by similar
concerns: we are currently exploiting the fact that `Shaded.class` is
accessible as a resource at compile-time, but this is not necessarily
true when we are compiling on the fly (again, similar concerns about
policies)

Notice also:

- this PR rationalizes and makes consistent the naming of the generated
classes with their contents; we now don't rely on classloading to
extract the name from the generated byte array, instead we always use
the key.
- in some cases the key in the map would not match the name of the class
in the byte array: e.g. Dispatcher, FunctionGroup, this PR fixes that
too
- notably, this PR also makes a tiny little change in the WasmAnalyzer
to ensure it compiles/run correctly on ART :)

---------

Signed-off-by: Edoardo Vacchi <evacchi@users.noreply.github.com>