Directory backed runtime compiler cache (#1077)
This expands on https://github.com/dylibso/chicory/pull/1041 and adds a
built in Directory Cache implementation.

When you use it with the runtime compiler:
```
          var cache = new DirectoryCache(Path.of("/cache"));
          var instance =
                Instance.builder(module)
                        .withMachineFactory(
                                MachineFactoryCompiler.builder(module).withCache(cache).compile())
                        .build();
```
It will create the files like:
```
/cache/.tmp/f-9979505440111212447.tmp
```

and then move the `/cache/.tmp/f-9979505440111212447.tmp` file in an
atomic operation to
`/cache/sha256/kr/gytkcm43c34ksqta8gmddw4ycfquc2g0qfifcpb-w.jar` (the
file path is based on the sha256 of the original wasm file).

It should be safe to use the same cache location by different processes.
If any error occurs while generating the code into the temp dir, the
temp file will get deleted. But if the process is killed before we can
clean up, the temp file will not get cleaned up. User may want manually
clean the /cache/.tmp directory when no process is writing to the cache.

we use 2 character prefix of the sha to avoid hitting directory scaling
issues with file systems. So if you have a VERY large cache all the
cache entries will be partitioned over 4096 dirs.

---------

Signed-off-by: Hiram Chirino <hiram@hiramchirino.com>
Co-authored-by: andreatp <andrea.peruffo1982@gmail.com>